+++ RogueSheep/Messaging/IMessage.cs
@@ -1,0 +1,8 @@
+namespace RogueSheep.Messaging
+{
+    public interface IMessage<T> where T : System.Enum
+    {
+        T MessageType { get; }
+        string Text { get; }
+    }
+}