+++ changelog.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **PCGRandom** class and helper to make the interface more System.Random-like
 - **IScheduler** interface and **RoundRobinScheduler** implementation
 - basic 2D geometry structs
+- Simple messaging system